requirements.txt
numpy>=1.23
PyYAML>=6.0
matplotlib>=3.7
Pillow>=10.0
